publish script now runs hugo!!

+++ publish.sh
@@ -1,1 +1,2 @@
-rsync --delete -va ~/git/myblog-hugo/public/ /home/user-data/www/stevehemingway.com
+#!/bin/sh
+hugo && rsync --delete -va ~/git/myblog-hugo/public/ /home/user-data/www/stevehemingway.com
